Changed the cover image of boefje/hello-katty to have an unique picture.
Instead of giving the address as a string to the Greeting model, it requires an IPv4 or IPv6 OOI to be able to have a reference to it.

Changed the risk of a KAT-GREETING finding from high to low to better display the realistic risk.

Removed the default_enabled property of the check_greeting bit as recommended by @noamblitz

Since the address of the Greeting model has been changed, it requires string conversion now.